Changed acceptable timezones
The RSS 2.0 spec states:
"All date-times in RSS conform to the Date and Time Specification of RFC 822, with the exception that the year may be

RFC 822 allows:
Time zone may be indicated in several ways. "UT" is Univer- sal Time (formerly called "Greenwich Mean Time"); "GMT" is

This allows these timezones.